+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Added quick retry for failed downloads on the download screen (ctrl + r)
+- Added caching for AllManga integration
 
 ## [1.3.0] - 2026-05-01